backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.1.0
openai>=1.68.0
pinecone>=6.0.0
pydantic>=2.10.0
python-multipart>=0.0.20
PyPDF2>=3.0.0
Pillow>=11.0.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
tenacity>=9.0.0
sse-starlette>=2.2.0
langchain-text-splitters>=0.3.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
